requirements.txt
appnope==0.1.0
article-website-crawling==0.1
backcall==0.1.0
beautifulsoup4==4.6.3
certifi==2018.10.15
chardet==3.0.4
cssselect==1.0.3
decorator==4.3.0
feedfinder2==0.0.4
feedparser==5.2.1
idna==2.7
ipykernel==5.1.0
ipython==7.1.1
ipython-genutils==0.2.0
jedi==0.13.1
jieba3k==0.35.1
jupyter-client==5.2.3
jupyter-core==4.4.0
lxml==4.2.5
newspaper3k==0.2.8
nltk==3.3
numpy==1.15.3
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
prompt-toolkit==2.0.6
ptyprocess==0.6.0
Pygments==2.2.0
python-dateutil==2.7.5
PyYAML==3.13
pyzmq==17.1.2
requests==2.20.0
requests-file==1.4.3
scikit-learn==0.20.0
scipy==1.1.0
six==1.11.0
tinysegmenter==0.3
tldextract==2.2.0
tornado==5.1.1
traitlets==4.3.2
urllib3==1.24
wcwidth==0.1.7
